+++ src/task-edit.js
@@ -2,9 +2,11 @@ import {WEEKDAYS, COLORLIST, MONTHLIST} from './export-const.js';
 import getUTCHours from './get-utc-hours.js';
 import getUTCMinutes from './get-utc-minutes.js';
 import createElement from './create-element.js';
+import Component from './component.js';
 
-export default class TaskEdit {
+export default class TaskEdit extends Component {
   constructor(task) {
+    super();
     this._color = task.color;
     this._date = task.date;
     this._repeatDays = task.repeatDays;
@@ -37,10 +39,6 @@ export default class TaskEdit {
 
   set onSubmit(fn) {
     this._onSubmit = fn;
-  }
-
-  get element() {
-    return this._element;
   }
 
   _onEditButtonClick() {
@@ -257,17 +255,6 @@ export default class TaskEdit {
     `;
 
     return cardContent.trim();
-  }
-
-  render() {
-    this._element = createElement(this.template);
-    this.bind();
-    return this._element;
-  }
-
-  unrender() {
-    this.unbind();
-    this._element = null;
   }
 
   bind() {
